Handle errors in `EventsRepositoryImpl.getEvent`
This commit updates `EventsRepositoryImpl.getEvent` to catch exceptions during event retrieval and return an empty `Event` object if an error occurs.

It also introduces an `ErrorLoadEvent` composable in `EventDetailsScreen` to display a message when an event cannot be loaded. The `Content` composable is now only displayed if the event ID is not empty.