Refactor: Optimize CityPlacesScreen with LazyColumn
This commit refactors the `CityPlacesScreen.kt` to use `LazyColumn` for displaying the list of city places.

Previously, the screen used a `Column` with `verticalScroll`. Replacing this with `LazyColumn` offers performance benefits by only composing and laying out the items that are currently visible on screen.

The main content display logic has been extracted into a new composable function named `Content` for better organization.